requirements_minimal.txt
# Minimal requirements for basic functionality
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=0.13.0
opencv-python>=4.8.0
librosa>=0.10.0
transformers>=4.30.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0
python-multipart>=0.0.6
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pillow>=10.0.0
requests>=2.31.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
soundfile>=0.12.0
scipy>=1.10.0
tqdm>=4.65.0
plotly>=5.15.0
seaborn>=0.12.0
reportlab>=4.0.0
openpyxl>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0
PyJWT>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
aiofiles>=23.0.0
imageio>=2.31.0